+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2023-06-17
+### Changed
+- Function signature of custom coercions to allow parsing more complex coercions types such as substr.
+
+### Added
+- Added new `_substr_[n:n]` COERCE identifier.
+
 ## [0.11.0] - 2023-06-10
 ### Added
 - Ability to register new, remove existing and replace existing COERCE types.
@@ -113,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/rust-playground/ksql/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/rust-playground/ksql/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/rust-playground/ksql/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/rust-playground/ksql/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/rust-playground/ksql/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/rust-playground/ksql/compare/v0.9.0...v0.9.1